Refactor Vulkan backend activation functions to use double precision
- Updated `dispatchActivation` and `dispatchActivationBackward` methods in `vulkan_backend.hpp` and `vulkan_ops_memory.cpp` to accept `double` parameters instead of `float` for better precision in activation functions.
- Adjusted kernel registration in `vulkan_kernel_registry.cpp` to pass `double` parameters directly, eliminating unnecessary casts.
- Enhanced comments to clarify the importance of maintaining precision for specific operations (LeakyReLU, ELU, Softplus).
- Modified MaskRCNN forward pass in `mask_rcnn.cpp` to ensure proper handling of class probabilities and labels, preventing out-of-range errors.
- Added support for testing with OneAPI CPU devices in `CMakeLists.txt` to improve compatibility across different hardware setups.
- Updated various unit tests to accommodate changes in memory management for Vulkan and CUDA backends, ensuring stability and performance across different data types and batch sizes.
- Introduced no-op stubs for CUDA configuration functions to maintain API consistency when CUDA is not built, preventing import errors in Python bindings.

+++ src/backends/vulkan/vulkan_backend.hpp
@@ -746,17 +746,23 @@ public:
     auto dispatchNMS(const Tensor& boxes, const Tensor& scores, float iou_threshold) -> Tensor;
 
     // Forward activation operations
+    //
+    // `param` is a double so a Float64-typed alpha (LeakyReLU/ELU) or
+    // beta (Softplus) survives the trip from the attrs map (where it is
+    // already a double) into the f64 push-constant struct without losing
+    // precision via an intermediate float cast. The f32/f16/bf16 paths
+    // implicitly narrow it back when filling their push constants.
     auto dispatchActivation(const std::string& op_name,
                            const Tensor& input,
                            uint32_t opcode,
-                           float param) -> Tensor;
+                           double param) -> Tensor;
 
     // Backward activation operations
     auto dispatchActivationBackward(const std::string& op_name,
                                      const Tensor& grad_output,
                                      const Tensor& input_or_output,
                                      uint32_t opcode,
-                                     float param) -> Tensor;
+                                     double param) -> Tensor;
     auto dispatchSwishBackward(const Tensor& grad_output, const Tensor& input) -> Tensor;
     auto dispatchRReLU(const Tensor& input, float lower, float upper, bool training) -> Tensor;
     auto dispatchRReLUBackward(const Tensor& grad_output, const Tensor& input, float slope) -> Tensor;
